Further refactoring automaton build logic - (Task #280)
---
Task #280: Improve time minimization of recovery automata
