+++ de.dlr.sc.virsat.fdir.core/src/de/dlr/sc/virsat/fdir/core/markov/AStateSpaceGenerator.java
@@ -1,5 +1,15 @@
+/*******************************************************************************
+ * Copyright (c) 2020 German Aerospace Center (DLR), Simulation and Software Technology, Germany.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
 package de.dlr.sc.virsat.fdir.core.markov;
 
+import java.util.Collections;
 import java.util.List;
 
 public abstract class AStateSpaceGenerator<S extends MarkovState> {
@@ -12,6 +22,17 @@ public abstract class AStateSpaceGenerator<S extends MarkovState> {
 	 */
 	public void init(MarkovAutomaton<S> targetMa) {
 		this.targetMa = targetMa;
+	}
+	
+	/**
+	 * Gets the first states that should be considered for state space generation.
+	 * By default this is only the initial state.
+	 * Special logic for generating special states only for the initial state can be put here.
+	 * @param initialState the initial state
+	 * @return the starting states
+	 */
+	public List<S> getStartingStates(S initialState) {
+		return Collections.singletonList(initialState);
 	}
 	
 	/**
