+++ de.dlr.sc.virsat.fdir.core/src/de/dlr/sc/virsat/fdir/core/markov/MarkovAutomatonBuilder.java
@@ -1,3 +1,12 @@
+/*******************************************************************************
+ * Copyright (c) 2020 German Aerospace Center (DLR), Simulation and Software Technology, Germany.
+ *
+ * This program and the accompanying materials are made available under the
+ * terms of the Eclipse Public License 2.0 which is available at
+ * http://www.eclipse.org/legal/epl-2.0.
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ *******************************************************************************/
 package de.dlr.sc.virsat.fdir.core.markov;
 
 import java.util.LinkedList;
@@ -5,6 +14,8 @@ import java.util.List;
 import java.util.Queue;
 
 public class MarkovAutomatonBuilder<S extends MarkovState> {
+	
+	private MarkovAutomatonBuildStatistics statistics = new MarkovAutomatonBuildStatistics();
 	
 	private AStateSpaceGenerator<S> stateSpaceGenerator;
 	private S initialState;
@@ -18,19 +29,28 @@ public class MarkovAutomatonBuilder<S extends MarkovState> {
 	 * @return the new markov automaton
 	 */
 	public MarkovAutomaton<S> build() {
+		statistics.time = System.currentTimeMillis();
+		
 		MarkovAutomaton<S> ma = new MarkovAutomaton<>();
 		stateSpaceGenerator.init(ma);
 		
 		initialState = stateSpaceGenerator.createInitialState();
 		
 		Queue<S> toProcess = new LinkedList<>();
-		toProcess.offer(initialState);
+		List<S> startingStates = stateSpaceGenerator.getStartingStates(initialState);
+		toProcess.addAll(startingStates);
 		
 		while (!toProcess.isEmpty()) {
 			S state = toProcess.poll();
 			List<S> generatedNewSuccs = stateSpaceGenerator.generateSuccs(state);
 			toProcess.addAll(generatedNewSuccs);
 		}
+		
+		System.out.println(ma.toDot());
+		
+		statistics.maxStates = ma.getStates().size();
+		statistics.maxTransitions = ma.getTransitions().size();
+		statistics.time = System.currentTimeMillis() - statistics.time;
 		
 		return ma;
 	}
@@ -41,5 +61,13 @@ public class MarkovAutomatonBuilder<S extends MarkovState> {
 	 */
 	public S getInitialState() {
 		return initialState;
+	}
+	
+	/**
+	 * Gets the internal statistics to the last call of the builder
+	 * @return the statistics of the last call of the builder
+	 */
+	public MarkovAutomatonBuildStatistics getStatistics() {
+		return statistics;
 	}
 }
