+++ de.dlr.sc.virsat.model.extension.fdir/src/de/dlr/sc/virsat/model/extension/fdir/evaluator/DFTEvaluator.java
@@ -23,7 +23,6 @@ import de.dlr.sc.virsat.fdir.core.metrics.FailLabelProvider;
 import de.dlr.sc.virsat.fdir.core.metrics.IBaseMetric;
 import de.dlr.sc.virsat.fdir.core.metrics.IDerivedMetric;
 import de.dlr.sc.virsat.fdir.core.metrics.IMetric;
-import de.dlr.sc.virsat.fdir.core.metrics.IQualitativeMetric;
 import de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.DFT2MAConverter;
 import de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.DFTState;
 import de.dlr.sc.virsat.model.extension.fdir.converter.dft2ma.DFTSymmetryChecker;
@@ -48,9 +47,8 @@ public class DFTEvaluator implements IFaultTreeEvaluator {
 	private DFTSemantics poSemantics;
 	
 	private MarkovAutomaton<DFTState> mc;
-	private RecoveryStrategy recoveryStrategy;
 	private IMarkovModelChecker markovModelChecker;
-	private DFT2MAConverter dft2MAConverter = new DFT2MAConverter();
+	private DFT2MAConverter dft2MaConverter = new DFT2MAConverter();
 	private Modularizer modularizer;
 	private DFTSymmetryChecker symmetryChecker;
 	private DFTMetricsComposer composer = new DFTMetricsComposer();
@@ -73,7 +71,7 @@ public class DFTEvaluator implements IFaultTreeEvaluator {
 
 	@Override
 	public void setRecoveryStrategy(RecoveryStrategy recoveryStrategy) {
-		this.recoveryStrategy = recoveryStrategy;
+		dft2MaConverter.getStateSpaceGenerator().setRecoveryStrategy(recoveryStrategy);
 	}
 
 	@Override
@@ -82,7 +80,7 @@ public class DFTEvaluator implements IFaultTreeEvaluator {
 		statistics.time = System.currentTimeMillis();
 		
 		FaultTreeHolder ftHolder = new FaultTreeHolder(root);
-		configureDFT2MAConverter(ftHolder, metrics, failableBasicEventsProvider);
+		dft2MaConverter.configure(ftHolder, chooseSemantics(ftHolder), metrics, failableBasicEventsProvider);
 		
 		DFTModularization modularization = getModularization(ftHolder, failableBasicEventsProvider);
 		
@@ -160,7 +158,7 @@ public class DFTEvaluator implements IFaultTreeEvaluator {
 	private DFTModularization getModularization(FaultTreeHolder ftHolder, FailableBasicEventsProvider failNodeProvider) {
 		boolean canModularize = modularizer != null 
 				&& ftHolder.getRoot() instanceof Fault
-				&& dft2MAConverter.getDftSemantics() != poSemantics
+				&& dft2MaConverter.getStateSpaceGenerator().getDftSemantics() != poSemantics
 				&& failNodeProvider == null;
 		
 		if (!canModularize) {
@@ -177,45 +175,6 @@ public class DFTEvaluator implements IFaultTreeEvaluator {
 	}
 	
 	/**
-	 * Configures the state space generator
-	 * @param ftHolder the fault tree to convert
-	 * @param metrics the metrics to evaluate
-	 * @param failableBasicEventsProvider the basic events provider
-	 */
-	private void configureDFT2MAConverter(FaultTreeHolder ftHolder, IMetric[] metrics, FailableBasicEventsProvider failableBasicEventsProvider) {
-		dft2MAConverter.setSemantics(chooseSemantics(ftHolder));
-		dft2MAConverter.setRecoveryStrategy(recoveryStrategy);
-		dft2MAConverter.getDftSemantics().setAllowsRepairEvents(!hasQualitativeMetric(metrics));
-		
-		if (dft2MAConverter.getDftSemantics() == poSemantics) {
-			dft2MAConverter.getStaticAnalysis().setSymmetryChecker(null);
-			dft2MAConverter.setAllowsDontCareFailing(false);
-		} else {
-			dft2MAConverter.getStaticAnalysis().setSymmetryChecker(new DFTSymmetryChecker());
-			dft2MAConverter.setAllowsDontCareFailing(true);
-		}
-		
-		if (failableBasicEventsProvider != null) {
-			dft2MAConverter.getStaticAnalysis().setSymmetryChecker(null);
-		}
-	}
-	
-	/**
-	 * Checks if there is a qualitative metric
-	 * @param metrics the metrics
-	 * @return true iff at least one metric is qualitative
-	 */
-	private boolean hasQualitativeMetric(IMetric[] metrics) {
-		for (IMetric metric : metrics) {
-			if (metric instanceof IQualitativeMetric) {
-				return true;
-			}
-		}
-		
-		return false;
-	}
-	
-	/**
 	 * Model checks a tree
 	 * @param root the root of the tree
 	 * @param subMonitor eclipse ui element for progress reporting
@@ -225,18 +184,13 @@ public class DFTEvaluator implements IFaultTreeEvaluator {
 	 * @return the result object containing the metrics
 	 */
 	private ModelCheckingResult modelCheck(FaultTreeNode root, SubMonitor subMonitor, IBaseMetric[] metrics, FailableBasicEventsProvider failableBasicEventsProvider, FailLabelProvider failLabelProvider) {
-		mc = dft2MAConverter.convert(root, failableBasicEventsProvider, failLabelProvider);
+		mc = dft2MaConverter.convert(root, failableBasicEventsProvider, failLabelProvider);
 		ModelCheckingResult result = markovModelChecker.checkModel(mc, subMonitor, metrics);
 			
-		statistics.stateSpaceGenerationStatistics.compose(dft2MAConverter.getStatistics());
+		statistics.maBuildStatistics.compose(dft2MaConverter.getMaBuilder().getStatistics());
 		statistics.modelCheckingStatistics.compose(markovModelChecker.getStatistics());	
 		
 		return result;
-	}
-	
-	@Override
-	public DFTEvaluationStatistics getStatistics() {
-		return statistics;
 	}
 	
 	/**
@@ -252,12 +206,9 @@ public class DFTEvaluator implements IFaultTreeEvaluator {
 		}
 	}
 	
-	/**
-	 * Gets the DFT2MA converter
-	 * @return the converter
-	 */
-	public DFT2MAConverter getDft2MAConverter() {
-		return dft2MAConverter;
+	@Override
+	public DFTEvaluationStatistics getStatistics() {
+		return statistics;
 	}
 	
 	/**
